backend/requirements.txt
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
ecdsa==0.19.1
fastapi==0.104.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==12.0
